fix(backend): correct security configuration and restore application startup

+++ backend/src/main/java/com/nutriguideai/dto/response/MealPlanHistoryResponse.java
@@ -14,5 +14,5 @@ import java.util.List;
 public class MealPlanHistoryResponse {
 
     private List<MealPlanDetailResponse> plans;
-    private long total;
+    private int total;
 }
